refactor(enrichment): extract class/module checks to _class_module.py
Extract 6 check functions, 11 helpers, and 5 constants (missing-attributes,
missing-typed-attributes, missing-examples, missing-cross-references,
prefer-fenced-code-blocks) into enrichment/_class_module.py. Deferred
import block re-exports 14 symbols for test and orchestrator access.

- Reduce enrichment/__init__.py from 2,259 to 1,450 lines (36% reduction)
- New _class_module.py at 857 lines with class/module enrichment checks
- Update module docstring to reflect sub-package structure
- Zero external API changes, zero test modifications

Closes #368
